Update NetTest_From_Web.py
Changed int to float for the think situation. This allows it to actually print an answer. I thought it was just wrong.

+++ NetTest_From_Web.py
@@ -73,4 +73,4 @@ if __name__ == "__main__":
 
     # Test the neural network with a new situation.
     print("Considering new situation [1, 0, 0] -> ?: ")
-    print(int(neural_network.think(array([1, 0, 0]))))
+    print(float(neural_network.think(array([1, 0, 0]))))
